% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanDupyStrings.R
\name{CleanDupyStrings}
\alias{CleanDupyStrings}
\title{CleanDupyStrings (operational function)}
\usage{
CleanDupyStrings(string_vec)
}
\arguments{
\item{string_vec}{A string}
}
\description{
This function is to clean the Player names field that contain accents.
}
