% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_all_squad_pages}
\alias{get_all_squad_pages}
\title{Get National Squads Plus U-2X}
\usage{
get_all_squad_pages(years = 2000:2020)
}
\arguments{
\item{years}{saison_id=years}
}
\description{
This function pulls data for all national squads plus their U-23, -21, etc squads
}
\examples{
get_all_squad_pages

}
\keyword{none}
