% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_mv_from_squad_page}
\alias{get_mv_from_squad_page}
\title{Get player data from national squad page}
\usage{
get_mv_from_squad_page(link_rel)
}
\arguments{
\item{link_rel}{The relative location of the squad page. www.transfermarkt.us/{link_rel}}
}
\description{
This function pulls data for all national squads plus their U-23, -21, etc squads
}
\examples{
get_mv_from_squad_page

}
\keyword{none}
