% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_national_squads}
\alias{get_national_squads}
\title{Get National Squads Function}
\usage{
get_national_squads(n)
}
\description{
This function pulls data for all national squads
}
\examples{
get_national_squads()

}
\keyword{none}
