% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_squad_list}
\alias{get_squad_list}
\title{Get Squad Players in List}
\usage{
get_squad_list(url, year)
}
\arguments{
\item{team_url}{year_ofsquad}
}
\description{
This function pulls data for all national squads
}
\examples{
get_squad_list()

}
\keyword{none}
