% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transfers.R
\name{get_transfers}
\alias{get_transfers}
\title{get_transfers}
\usage{
get_transfers(url, season = NULL, include_end_of_loans = TRUE,
  include_retired = TRUE, include_without_club = TRUE)
}
\arguments{
\item{url}{The url for the relevant page to scrape. Example: 'https://www.transfermarkt.co.uk/premier-league/sommertransfers/wettbewerb/GB1'}

\item{season}{The season year.}

\item{include_end_of_loans}{Boolean value for whether to include or not players returning from loans in the results. Defaults to TRUE.}

\item{include_retired}{Boolean value for whether to include or not players that retired. Defaults to TRUE.}

\item{include_without_club}{Boolean value for whether to include or not players without club in the results. Defaults to TRUE.}
}
\value{
A data-frame where each row consists of a player transfer. The columns returned are From, To, Player, Age, Position, Position_Code, Market_Value, Market_Value_new, Fee, Fee_new, Transfer_Type, Transfer_Direction, Club, Club_Icon, League, Season_Scrapped, Date_Scrapped.
}
\description{
This function scrappes a transfermarkt webpage to collect a data-frame table with all the transfers for a particular league, season and transfer window.
}
\examples{
\dontrun{
tranfers_premierLeague = get_transfers(url = "https://www.transfermarkt.co.uk/premier-league/sommertransfers/wettbewerb/GB1",
                                     season = "2018",
                                     include_end_of_loans = TRUE,
                                     include_retired = TRUE,
                                     include_without_club = TRUE
                                     )


}
}
