% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodesAndEdges.R
\name{NodesAndEdges}
\alias{NodesAndEdges}
\title{NodesAndEdges (operational function)}
\usage{
NodesAndEdges(df, value_bins = 3)
}
\arguments{
\item{df}{A df}

\item{value_bins}{The number of bins. Defaults to 3.}
}
\description{
This function is used to format the transfer data into nodes and edges format to be supported by the visNetwork library.
}
