% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_market_values.R
\name{pull_national_squads}
\alias{pull_national_squads}
\title{Get National Squads Function}
\usage{
pull_national_squads()
}
\arguments{
\item{none}{ad}
}
\description{
This function pulls data for all national squads
}
\examples{
pull_national_squads()

}
\keyword{none}
