% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transfermarktR_packages}
\alias{transfermarktR_packages}
\title{List all packages in the transfermarktR}
\usage{
transfermarktR_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include transfermarktR in the list?}
}
\description{
List all packages in the transfermarktR
}
\examples{
transfermarktR_packages()
}
