% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transfers_NetworkPlot.R
\name{Transfers_NetworkPlot}
\alias{Transfers_NetworkPlot}
\title{Transfers_NetworkPlot}
\usage{
Transfers_NetworkPlot(transfers_df, transfer_fee_bins = 3)
}
\arguments{
\item{transfers_df}{A data-frame returned from get_transfers function.}

\item{transfer_fee_bins}{The number of bins to define from the Fee of the transfer. This is used to define the color of the links on the visualisation illustrating the value of a transaction. Defaults to 3.}
}
\value{
A visNetwork graph.
}
\description{
This function creates a Network plot (visNetwork) with the transfer data scrapped before.
}
