% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_market_values.R
\name{get_all_player_mv}
\alias{get_all_player_mv}
\title{Get National Squads Function}
\usage{
get_all_player_mv()
}
\arguments{
\item{none}{ad}
}
\description{
This function pulls data for all national squads
}
\examples{
get_all_player_mv()

}
\keyword{none}
